feat: relate a task with an user

+++ src/tasks/tasks.service.ts
@@ -1,6 +1,7 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 
 import { InjectRepository } from '@nestjs/typeorm';
+import User from 'src/auth/user.entity';
 import GetTasksFilterDTO from './dto/get-tasks-filter.dto';
 import TaskRepository from './task.repository';
 import Task from './task.entity';
@@ -14,14 +15,21 @@ export default class TasksService {
     private taskRepository: TaskRepository,
   ) {}
 
-  async getTasks(filterDTO: GetTasksFilterDTO): Promise<Task[]> {
-    const tasks = await this.taskRepository.getTasks(filterDTO);
+  async getTasks(filterDTO: GetTasksFilterDTO, user: User): Promise<Task[]> {
+    const tasks = await this.taskRepository.getTasks(filterDTO, user);
 
     return tasks;
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const task = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const { id: userId } = user;
+
+    const task = await this.taskRepository.findOne({
+      where: {
+        id,
+        userId,
+      },
+    });
 
     if (!task) {
       throw new NotFoundException(`Task with ID ${id} not found`);
@@ -30,22 +38,28 @@ export default class TasksService {
     return task;
   }
 
-  async createTask(createTaskDTO: CreateTaskDTO): Promise<Task> {
-    const task = await this.taskRepository.createTask(createTaskDTO);
+  async createTask(createTaskDTO: CreateTaskDTO, user: User): Promise<Task> {
+    const task = await this.taskRepository.createTask(createTaskDTO, user);
 
     return task;
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const { affected } = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<void> {
+    const { id: userId } = user;
+
+    const { affected } = await this.taskRepository.delete({ id, userId });
 
     if (affected === 0) {
       throw new NotFoundException(`Task with ID ${id} not found`);
     }
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getTaskById(id, user);
 
     task.status = status;
